requirements.txt
asgiref==3.6.0
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
generics==6.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.3
